+++ src/game/server/explode.cpp
@@ -106,7 +106,7 @@ public:
 	void Precache( void );
 	void Spawn( );
 	void Smoke ( void );
-	void SetCustomDamageType( int iType ) { m_iCustomDamageType = iType; }
+	void SetCustomDamageType( uint64 iType ) { m_iCustomDamageType = iType; }
 	bool KeyValue( const char *szKeyName, const char *szValue );
 
 	int DrawDebugTextOverlays(void);
@@ -124,7 +124,7 @@ public:
 	string_t m_iszFireballSprite;
 	short m_sFireballSprite;
 	EHANDLE m_hInflictor;
-	int m_iCustomDamageType;
+	uint64 m_iCustomDamageType;
 
 	// passed along to the RadiusDamage call
 	int m_iClassIgnore;
@@ -337,7 +337,7 @@ void CEnvExplosion::InputExplode( inputdata_t &inputdata )
 		CBaseEntity *pAttacker = GetOwnerEntity() ? GetOwnerEntity() : this;
 
 		// Only calculate damage type if we didn't get a custom one passed in
-		int iDamageType = m_iCustomDamageType;
+		uint64 iDamageType = m_iCustomDamageType;
 		if ( iDamageType == -1 )
 		{
 			iDamageType = HasSpawnFlags( SF_ENVEXPLOSION_GENERIC_DAMAGE ) ? DMG_GENERIC : DMG_BLAST;
@@ -400,7 +400,7 @@ void CEnvExplosion::Smoke( void )
 
 // HACKHACK -- create one of these and fake a keyvalue to get the right explosion setup
 void ExplosionCreate( const Vector &center, const QAngle &angles, 
-	CBaseEntity *pOwner, int magnitude, int radius, int nSpawnFlags, float flExplosionForce, CBaseEntity *pInflictor, int iCustomDamageType,
+	CBaseEntity *pOwner, int magnitude, int radius, int nSpawnFlags, float flExplosionForce, CBaseEntity *pInflictor, uint64 iCustomDamageType,
 	const EHANDLE *ignoredEntity , Class_T ignoredClass )
 {
 	char			buf[128];
@@ -442,7 +442,7 @@ void ExplosionCreate( const Vector &center, const QAngle &angles,
 
 
 void ExplosionCreate( const Vector &center, const QAngle &angles, 
-	CBaseEntity *pOwner, int magnitude, int radius, bool doDamage, float flExplosionForce, bool bSurfaceOnly, bool bSilent, int iCustomDamageType )
+	CBaseEntity *pOwner, int magnitude, int radius, bool doDamage, float flExplosionForce, bool bSurfaceOnly, bool bSilent, uint64 iCustomDamageType )
 {
 	// For E3, no sparks
 	int nFlags = SF_ENVEXPLOSION_NOSPARKS | SF_ENVEXPLOSION_NODLIGHTS | SF_ENVEXPLOSION_NOSMOKE;
@@ -468,7 +468,7 @@ void ExplosionCreate( const Vector &center, const QAngle &angles,
 void ExplosionCreate( const Vector &center, const QAngle &angles, 
 					 CBaseEntity *pOwner, int magnitude, int radius, bool doDamage, 
 					 const EHANDLE *ignoredEntity, Class_T ignoredClass,
-					 float flExplosionForce , bool bSurfaceOnly , bool bSilent , int iCustomDamageType )
+					 float flExplosionForce , bool bSurfaceOnly , bool bSilent , uint64 iCustomDamageType )
 {
 	// For E3, no sparks
 	int nFlags = SF_ENVEXPLOSION_NOSPARKS | SF_ENVEXPLOSION_NODLIGHTS | SF_ENVEXPLOSION_NOSMOKE;
