+++ src/game/shared/ammodef.h
@@ -12,6 +12,8 @@
 
 #pragma once
 
+#include "tier0/platform.h"
+
 #ifdef GAME_DLL
 class CBaseCombatCharacter;
 typedef CBaseCombatCharacter CSharedBaseCombatCharacter;
@@ -31,7 +33,7 @@ class ConVar;
 struct Ammo_t 
 {
 	char 				*pName;
-	int					nDamageType;
+	uint64					nDamageType;
 	int					eTracerType;
 	float				physicsForceImpulse;
 	int					nMinSplashSize;
@@ -90,15 +92,15 @@ public:
 	int					NPCDamage(int nAmmoIndex);
 	int					MaxCarry(int nAmmoIndex, const CSharedBaseCombatCharacter *owner);
 	bool				CanCarryInfiniteAmmo(int nAmmoIndex);
-	int					DamageType(int nAmmoIndex);
+	uint64					DamageType(int nAmmoIndex);
 	int					TracerType(int nAmmoIndex);
 	float				DamageForce(int nAmmoIndex);
 	int					MinSplashSize(int nAmmoIndex);
 	int					MaxSplashSize(int nAmmoIndex);
 	int					Flags(int nAmmoIndex);
 
-	void				AddAmmoType(char const* name, int damageType, int tracerType, int plr_dmg, int npc_dmg, int carry, float physicsForceImpulse, int nFlags, int minSplashSize = 4, int maxSplashSize = 8 );
-	void				AddAmmoType(char const* name, int damageType, int tracerType, char const* plr_cvar, char const* npc_var, char const* carry_cvar, float physicsForceImpulse, int nFlags, int minSplashSize = 4, int maxSplashSize = 8 );
+	void				AddAmmoType(char const* name, uint64 damageType, int tracerType, int plr_dmg, int npc_dmg, int carry, float physicsForceImpulse, int nFlags, int minSplashSize = 4, int maxSplashSize = 8 );
+	void				AddAmmoType(char const* name, uint64 damageType, int tracerType, char const* plr_cvar, char const* npc_var, char const* carry_cvar, float physicsForceImpulse, int nFlags, int minSplashSize = 4, int maxSplashSize = 8 );
 
 	int					NumAmmoTypes() { return m_nAmmoIndex; }
 	int					GetNumAmmoTypes() { return m_nAmmoIndex; }
@@ -107,7 +109,7 @@ public:
 	virtual ~CAmmoDef( void );
 
 private:
-	bool				AddAmmoType(char const* name, int damageType, int tracerType, int nFlags, int minSplashSize, int maxSplashSize );
+	bool				AddAmmoType(char const* name, uint64 damageType, int tracerType, int nFlags, int minSplashSize, int maxSplashSize );
 };
