+++ src/game/shared/gamerules.h
@@ -241,17 +241,17 @@ public:
 
 	// Damage Queries - these need to be implemented by the various subclasses (single-player, multi-player, etc).
 	// The queries represent queries against damage types and properties.
-	virtual bool	Damage_IsTimeBased( int iDmgType );			// Damage types that are time-based.
-	virtual bool	Damage_ShouldGibCorpse( int iDmgType );		// Damage types that gib the corpse.
-	virtual bool	Damage_ShowOnHUD( int iDmgType );			// Damage types that have client HUD art.
-	virtual bool	Damage_NoPhysicsForce( int iDmgType );		// Damage types that don't have to supply a physics force & position.
-	virtual bool	Damage_ShouldNotBleed( int iDmgType );		// Damage types that don't make the player bleed.
+	virtual bool	Damage_IsTimeBased( uint64 iDmgType );			// Damage types that are time-based.
+	virtual bool	Damage_ShouldGibCorpse( uint64 iDmgType );		// Damage types that gib the corpse.
+	virtual bool	Damage_ShowOnHUD( uint64 iDmgType );			// Damage types that have client HUD art.
+	virtual bool	Damage_NoPhysicsForce( uint64 iDmgType );		// Damage types that don't have to supply a physics force & position.
+	virtual bool	Damage_ShouldNotBleed( uint64 iDmgType );		// Damage types that don't make the player bleed.
 	//Temp: These will go away once DamageTypes become enums.
-	virtual int		Damage_GetTimeBased( void );				// Actual bit-fields.
-	virtual int		Damage_GetShouldGibCorpse( void );
-	virtual int		Damage_GetShowOnHud( void );					
-	virtual int		Damage_GetNoPhysicsForce( void );
-	virtual int		Damage_GetShouldNotBleed( void );
+	virtual uint64		Damage_GetTimeBased( void );				// Actual bit-fields.
+	virtual uint64		Damage_GetShouldGibCorpse( void );
+	virtual uint64		Damage_GetShowOnHud( void );					
+	virtual uint64		Damage_GetNoPhysicsForce( void );
+	virtual uint64		Damage_GetShouldNotBleed( void );
 
 	void LoadVoiceCommandScript( void );
